backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pillow==11.0.0
torch
diffusers==0.31.0
transformers==4.47.1
accelerate==1.2.1
safetensors==0.4.5
